+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 -   Remove dependency to config module. [#317](https://github.com/atomist/automation-client-ts/issues/317)
+-   `Project` no longer extends `AbstractScriptedFlushable`. This is no longer necessary given async/await
 
 ## [0.17.1](https://github.com/atomist/automation-client-ts/compare/0.17.0...0.17.1) - 2018-06-04